requirements.txt
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-cloud-bigquery>=3.11.0
db-dtypes>=1.2.0
youtube-transcript-api>=0.6.2
requests>=2.31.0
streamlit>=1.35.0
plotly>=5.20.0
pandas>=2.2.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0